update
+fix dont deal damage
+fix pasuing function
+add a chair object concept

+++ Chef Strikes Back/Assets/Scripts/Player/Player.cs
@@ -113,7 +113,7 @@ public class Player : MonoBehaviour
         actionState.FixedUpdate();
     }
 
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         currentHealth -= damageAmount;
